Relative mouse position
Edited the pageToLocalCoords function so that it actually appears to
work at the moment, at least in chrome.

Haven't deleted the code I changed, only commented it out, just in case.

Also I made the debug console scroll to the bottom every time a new
status message occurs, which makes it much easier to use for things like
mouse positions and button clicking.